Add code to fetch data and present
why: The wanted result is a combination of the result from two API
endpoints, with a common identifier. These has to be stored and combined
before presentation of the result is possible.

How: Two dictionaries is created from the endpoints, and then merged to
a common list of lists, each containing information about one station.